ssid and channel should be nullable
When registering a phone number, there's two steps: creating the
database record, and sending the number over to twilio (with a
channel: "call" or "sms" in this case) to get the sid.

For some reason I think the right-for-now thing to do is to break this
in to two functions: create the record locally first, and do the
twilio stuff (record get the sid and channel) later.