adding Lenny.Twilio / Lenny.TwilioImpl
This is all for the sake of testability.  Twilio is a behavior which
also defines functions that delegate to a configurable implementation,
which defaults to TwilioImpl.

The big idea is that, when writing tests, we can reconfigure the
application environment to set the default implementation to a
mock/stub with the help of the Mox library.

I'm not a big fan of mocks / stubs / etc., but I'm hoping that this
enables some useful automated tests.