showing in-progress call when changing phone number
In other words, if there's a call in progress for +13125551234, and
you change and verify your phone number from something else to
+13125551234, then you should immediately see that call in progress.

However this seems like an ugly path.  What happens when a call comes
in for the old number?  We need to ensure that channels for the old
number are unsubscribed from, channels for the new number are
subscribed to, etc.,

The next approach I want to take is break LennyLive apart in a way
such that the LiveView doesn't have to worry about numbers changing.