separating CallLive into CallLive and WaitLive
WaitLive waits for {:call_started, sid} events on the call:#{phone}
channel.  When in gets one, it redirects to /call/#{sid} which listens
to events (namely :call_ended) on call:#{sid}.  /call/#{sid} does not
require authentication; the idea is that the sids are not guessable,
and this enables a "try this out by calling this number, and we'll
text you a link" demo.