both tracks (inbound, outbound) seem to be playing smoothly now
I have no idea why this is working.  But it seems you have to refresh
"callStart" every so often, as there's some kind of drift that
eventually causes the audio to get choppy.

And it seems like the inbound track gets choppier much faster, so it
needs to be refreshed more frequently.