passing pending/verified numbers to PhoneNumberLive via "session"
I think I finally understand this now.  According to the documentation
of live_render/3's :session option:

  :session - a map of binary keys with extra session data to be
  serialized and sent to the client. All session data currently in the
  connection is automatically available in LiveViews. You can use this
  option to provide extra data. Remember all session data is
  serialized and sent to the client, so you should always keep the
  data in the session to a minimum. For example, instead of storing a
  User struct, you should store the "user_id" and load the User when
  the LiveView mounts.

I've always been super confused by this.  The session is a singular
data store per-client, is it not?  Would two components setting the
same key somehow "clobber" eachother?

Turns out no.  And digging through the source code, it sort of makes
sense.  The extra data passed via :session is local to the resulting
live view process.  The part about it being serialized and sent to the
client is referring to how that data is serialized and ends up in the
data-phx-session attribute of the element in the DOM containing the
rendered live view.  Why does that need to happen?  I'm not quite
sure, but I'm guessing it has something to do with being able to
restore the server processes and their state without doing a full page
reload if the server processes die or if the websocket connection gets
interrupted.