reverting the previous two commits and trying something else
Keeping PhoneNumberLive -- it's easy to test using live_isolated.  No
need to pass it the pending number, as it can just get it itself and
neither of the hosting pages (CallsLive during new user registration,
settings) actually need it.  Not outputing instructions for the phone
number input as it varies between the two host pages.  Instructions
for entering the code are the same, so keeping that in
PhoneNumberLive.