almost working the way I want it.
We're back once again to having the thing that lets users enter and
verify their phone number being a full-on LiveView.  The one issue now
is that on CallsLive, we want the "Get started by verifying a phone
number" message to disappear once the users submits a phone number in
PhoneNumberLive, but we can't without some kind of pubsub nonsense.

So, advantages to LiveView:

  - easy to test in isolation

Disadvantages:

  - can't pass slots
  - awkward passing parameters (gives "session" a dual/mixed meaning)
  - awkward handling of the layout

Advantages of LiveComponent

  - slots

Disadvantages

  - hard to test in isolation
  - assigns not available during mount (why?)
  - in this case, awkwardness from taking "edit phone" out of
    UserSessionController